services/user-service/requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
python-jose
email-validator
passlib==1.7.4
bcrypt==4.0.1
pydantic-settings
